VFB-264 Refactor component definitions into separate files
VFB-264 remove faulty cypress test
