+++ src/app/info/AdminManagerDependent.tsx
@@ -1,0 +1,16 @@
+"use client";
+
+import { useContext } from "react";
+import { RoleUpdateContext } from "@/app/roles";
+
+interface RoleProps {
+    children?: React.ReactNode;
+}
+
+const AdminManagerDependent: React.FC<RoleProps> = (props) => {
+    const { role } = useContext(RoleUpdateContext);
+
+    return <>{(role === "admin" || role === "manager") && props.children}</>;
+};
+
+export default AdminManagerDependent;
