+++ src/app/info/StyleComponents.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Paper, TextField } from "@mui/material";
+import { Paper } from "@mui/material";
 import styled from "styled-components";
 
 export const WikiItemPositioner = styled.div`
